remove code that won't be needed for reading, move code that will be needed for writing to System.Windows.Forms where it's used (System.Windows.Forms compiles)
