+++ src/System.Windows.Forms/src/System/Windows/Forms/BinaryFormat/BinaryArrayType.cs
