+++ src/System.Windows.Forms/src/System/Windows/Forms/BinaryFormat/BinaryFormatWriterScope.cs
@@ -1,7 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Runtime.Serialization.BinaryFormat;
 using System.Text;
 
 namespace System.Windows.Forms.BinaryFormat;
@@ -13,19 +12,14 @@ internal readonly ref struct BinaryFormatWriterScope
     public BinaryFormatWriterScope(Stream stream)
     {
         Writer = new BinaryWriter(stream, Encoding.UTF8, leaveOpen: true);
-        // SerializationHeader
-        Writer.Write((byte)RecordType.SerializedStreamHeader);
-        Writer.Write(1); // root ID
-        Writer.Write(1); // header ID
-        Writer.Write(1); // major version
-        Writer.Write(0); // minor version
+        SerializationHeader.Default.Write(Writer);
     }
 
     public static implicit operator BinaryWriter(in BinaryFormatWriterScope scope) => scope.Writer;
 
     public void Dispose()
     {
-        Writer.Write((byte)RecordType.MessageEnd);
+        MessageEnd.Instance.Write(Writer);
         Writer.Dispose();
     }
 }
