+++ src/System.Windows.Forms/src/System/Windows/Forms/BinaryFormat/BinaryObjectString.cs
@@ -13,7 +13,7 @@ namespace System.Windows.Forms.BinaryFormat;
 ///   </see>
 ///  </para>
 /// </remarks>
-internal sealed class BinaryObjectString : IRecord<BinaryObjectString>, IBinaryFormatParseable<BinaryObjectString>
+internal sealed class BinaryObjectString : IRecord<BinaryObjectString>
 {
     public Id ObjectId { get; }
     public string Value { get; }
@@ -26,9 +26,6 @@ internal sealed class BinaryObjectString : IRecord<BinaryObjectString>, IBinaryF
         ObjectId = objectId;
         Value = value;
     }
-
-    static BinaryObjectString IBinaryFormatParseable<BinaryObjectString>.Parse(BinaryFormattedObject.IParseState state) =>
-        new(state.Reader.ReadInt32(), state.Reader.ReadString());
 
     public void Write(BinaryWriter writer)
     {
