+++ src/System.Windows.Forms/src/System/Windows/Forms/BinaryFormat/ClassInfo.cs
@@ -26,20 +26,6 @@ internal class ClassInfo : IBinaryWriteable
         MemberNames = memberNames;
     }
 
-    public static ClassInfo Parse(BinaryReader reader, out Count memberCount)
-    {
-        Id objectId = reader.ReadInt32();
-        string name = reader.ReadString();
-        memberCount = reader.ReadInt32();
-        ArrayBuilder<string> memberNames = new(memberCount);
-        for (int i = 0; i < memberCount; i++)
-        {
-            memberNames.Add(reader.ReadString());
-        }
-
-        return new(objectId, name, memberNames.ToArray());
-    }
-
     public void Write(BinaryWriter writer)
     {
         writer.Write(ObjectId);
