+++ src/System.Windows.Forms/src/System/Windows/Forms/BinaryFormat/IBinaryArray.cs
