+++ src/System.Windows.Forms/src/System/Windows/Forms/BinaryFormat/MessageEnd.cs
@@ -6,16 +6,13 @@ namespace System.Windows.Forms.BinaryFormat;
 /// <summary>
 ///  Record that marks the end of the binary format stream.
 /// </summary>
-internal sealed class MessageEnd : IRecord<MessageEnd>, IBinaryFormatParseable<MessageEnd>
+internal sealed class MessageEnd : IRecord<MessageEnd>
 {
     public static MessageEnd Instance { get; } = new();
 
     private MessageEnd() { }
 
     public static RecordType RecordType => RecordType.MessageEnd;
-
-    static MessageEnd IBinaryFormatParseable<MessageEnd>.Parse(
-        BinaryFormattedObject.IParseState state) => Instance;
 
     public void Write(BinaryWriter writer) => writer.Write((byte)RecordType);
 }
