+++ src/System.Windows.Forms/src/System/Windows/Forms/BinaryFormat/NullRecord.cs
